# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.createSubmissionRequest:
    properties:
      code:
        type: string
      language:
        type: string
    required:
    - code
    - language
    type: object
...
  api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.loginResponse:
    properties:
      expiresInSeconds:
        type: integer
      sessionToken:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
...
  judge.Result:
    properties:
      output:
        type: string
      status:
        $ref: '#/definitions/judge.Status'
      submission_id:
        type: string
    type: object
  judge.Status:
    enum:
    - queued
    - running
    - passed
    - failed
    type: string
    x-enum-varnames:
    - StatusQueued
    - StatusRunning
    - StatusPassed
    - StatusFailed
...
  team.Team:
    properties:
      createdAt:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/user.User'
        type: array
      name:
        type: string
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      role:
        type: string
      teamId:
        type: string
      teamName:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Algorithmic contest platform REST API & judge engine.
  title: MiniAlgothon API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, creating a server-side
        session.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Delete current server-side session token.
      responses:
        "244":
          description: No Content
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Auth
  /api/v1/me:
    get:
      description: Fetch authenticated user profile details and team info.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - Auth
...
  /api/v1/submissions:
    post:
      consumes:
      - application/json
      description: Queue code submission for official contest judging.
      parameters:
      - description: Submission payload
        in: body
        name: submission
        required: true
        schema:
          $ref: '#/definitions/api.createSubmissionRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit Code
      tags:
      - Submissions
  /api/v1/submissions/{id}:
    get:
      description: Fetch evaluation verdict and test results by submission ID.
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/judge.Result'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Submission Result
      tags:
      - Submissions
  /api/v1/teams:
    get:
      description: Fetch all teams and their assigned member accounts.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/team.Team'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Teams
      tags:
      - Teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"